requirements.txt
anthropic>=0.20.0
requests>=2.31.0
beautifulsoup4>=4.12.0
duckduckgo-search>=3.9.0
reportlab>=4.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
lxml>=4.9.0
